change the computation of the max cache size
Before this the number of bytes which are configured in openfire are set es megabyte.
Also change the API to make it able to set the max cache size value as long.

+++ src/java/org/jivesoftware/openfire/plugin/util/cache/ClusteredCache.java
@@ -148,9 +148,9 @@ public class ClusteredCache<K extends Serializable, V extends Serializable> impl
     }
 
     @Override
-    public int getCacheSize() {
+    public long getCacheSize() {
         final LocalMapStats stats = map.getLocalMapStats();
-        return (int) (stats.getOwnedEntryMemoryCost() + stats.getBackupEntryMemoryCost());
+        return stats.getOwnedEntryMemoryCost() + stats.getBackupEntryMemoryCost();
     }
 
     @Override
@@ -159,7 +159,7 @@ public class ClusteredCache<K extends Serializable, V extends Serializable> impl
     }
 
     @Override
-    public void setMaxCacheSize(final int maxSize) {
+    public void setMaxCacheSize(final long maxSize) {
         CacheFactory.setMaxSizeProperty(getName(), maxSize);
     }
